bigdata_platform/requirements.txt
pyspark
networkx
pycryptodome
findspark
pymsyql